requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.0
Flask-JWT-Extended==4.7.1
python-dotenv==1.1.0
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
eventlet==0.39.1
python-engineio==4.12.0
python-socketio==5.13.0 
